Separate core/ui in TypeMismatchSubProcessor

Fix failing tests
